[1407] Adds logic functions for resource groups.
Adds the following resource group action functions.

- resource_group_list
Returns details (label and id) of resource groups in the package provided in the id part of the call. User must have package_show permission on the containing package.

- resource_group_show
Returns resource group details, and full resources (dictized) for the given resource group if the user has package_show permission on the containing package.